refactor: move attach/detach profile
